Fixed bug when overshooting block when moving
The reason it wasn't work is the body had no height, so moving
off the block wasn't actually a `mousemove` on the body,
which is what we were listening to. Moving back onto the block
was back in the body, so it started moving again. This makes
the body the same height as the width, which should work
for any demos. A real app won't have this problem, so it doesn't
meaningfully impact what I'm trying to say in the presentation.